+++ Packages/UGF.Module.Controllers/Runtime/IController.cs
@@ -1,10 +1,9 @@
 ﻿using UGF.Application.Runtime;
-using UGF.Description.Runtime;
 using UGF.Initialize.Runtime;
 
 namespace UGF.Module.Controllers.Runtime
 {
-    public interface IController : IInitialize, IDescribed
+    public interface IController : IInitialize
     {
         IApplication Application { get; }
     }