+++ Packages/UGF.Module.Controllers/Runtime/IControllerDescribed.cs
@@ -1,0 +1,8 @@
+﻿using UGF.Description.Runtime;
+
+namespace UGF.Module.Controllers.Runtime
+{
+    public interface IControllerDescribed : IController, IDescribed
+    {
+    }
+}